Better Cache Detection and Pointer cache added
Some games that loops a same pointer can causes a high memory usage, to prevent this I added a new feature to cache pointers, This can cause missmatch in some games that use the same buffer to all strings and by this reason enable this feature results in a warning in  the log. by default disabled.